+++ othoz/paragraph/tests/test_session.py
@@ -2,14 +2,14 @@ import pytest
 
 from othoz.paragraph.types import Variable
 from othoz.paragraph.session import traverse_fw, traverse_bw, evaluate, solve_requirements, apply
-from othoz.paragraph.tests.test_types import TestReq, mock_op
+from othoz.paragraph.tests.test_types import MockReq, mock_op
 
 
 @pytest.fixture
 def graph():
     graph = lambda: None  # noqa: E731
     op1 = mock_op("op0")
-    graph.input = Variable()
+    graph.input = Variable("input")
     output0 = op1(arg=graph.input)
 
     op2 = mock_op("op1")
@@ -21,7 +21,8 @@ def graph():
 
 
 class TestForwardGenerator:
-    def test_generated_values(self, graph):
+    @staticmethod
+    def test_generated_values(graph):
         items = list(traverse_fw(graph.output))
         expected = [graph.input] + graph.output
 
@@ -29,7 +30,8 @@ class TestForwardGenerator:
 
 
 class TestBackwardGenerator:
-    def test_generated_value_wo_boundary(self, graph):
+    @staticmethod
+    def test_generated_value_wo_boundary(graph):
         items = list(traverse_bw(graph.output))
         expected = [graph.output[1], graph.output[0], graph.input]
 
@@ -39,18 +41,20 @@ class TestBackwardGenerator:
 @pytest.mark.parametrize("max_workers", [pytest.param(1, id="Single thread"),
                                          pytest.param(50, id="Multi-threaded")])
 class TestEvaluate:
-    def test_variable_evaluation_is_correct(self, graph, max_workers):
+    @staticmethod
+    def test_variable_evaluation_is_correct(graph, max_workers):
         res = evaluate(graph.output, args={graph.input: "Input value"}, max_workers=max_workers)
 
         assert res[0] == "op0 return value"
         assert res[1] == "op1 return value"
 
-        graph.output[0].func.func._run.assert_called_once_with(arg="Input value")
-        graph.output[1].func.func._run.assert_called_once_with(arg0="Input value", arg1="op0 return value")
+        graph.output[0].op._run.assert_called_once_with(arg="Input value")
+        graph.output[1].op._run.assert_called_once_with(arg0="Input value", arg1="op0 return value")
 
-    def test_evaluation_is_lazy(self, graph, max_workers):
+    @staticmethod
+    def test_evaluation_is_lazy(graph, max_workers):
         res = evaluate([graph.output[0]], args={graph.input: "Input value"}, max_workers=max_workers)
-        operation = graph.output[1].func.func
+        operation = graph.output[1].op
 
         assert res[0] == "op0 return value"
         assert not operation._run.called
@@ -59,27 +63,30 @@ class TestEvaluate:
 @pytest.mark.parametrize("max_workers", [pytest.param(1, id="Single thread"),
                                          pytest.param(50, id="Multi-threaded")])
 class TestApply:
-    def test_apply_returns_correct_number_of_results(self, graph, max_workers):
+    @staticmethod
+    def test_apply_returns_correct_number_of_results(graph, max_workers):
         res = list(apply(graph.output, args={}, iter_args=[{graph.input: "Input value"}] * 5, max_workers=max_workers))
 
         assert len(res) == 5
 
-    def test_apply_raises_on_overwriting_an_input_variable(self, graph, max_workers):
+    @staticmethod
+    def test_apply_raises_on_overwriting_an_input_variable(graph, max_workers):
         with pytest.raises(ValueError):
             list(apply(graph.output, args={graph.input: "Input value"}, iter_args=[{graph.input: "Input value"}] * 5, max_workers=max_workers))
 
 
 class TestRequirementSolving:
-    def test_req_update_func_called(self):
-        operation = mock_op("")
+    @staticmethod
+    def test_req_update_func_called():
+        operation = mock_op()
 
-        var = Variable()
+        var = Variable("input")
         res = operation(a=1, b=var)
 
-        output_req = TestReq("Output requirement")
+        output_req = MockReq("Output requirement")
         reqs = solve_requirements(output_requirements={res: output_req})
 
-        expected = {res: output_req, var: TestReq("Arg requirement")}
+        expected = {res: output_req, var: MockReq("Arg requirement")}
 
         assert reqs == expected
-        res.arg_requirements_func.assert_called_once_with(output_req, "b")
+        res.op.arg_requirements.assert_called_once_with(output_req, "b")